chore: move make_url into utils.py
